Refactor logging utils in a dedicated module
